Backend/requirements.txt
asgiref==3.5.2
astroid==2.11.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.0.4
django-cors-headers==3.12.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
